+++ chopper_generator/lib/src/generator.dart
@@ -2,6 +2,7 @@
 import 'dart:async';
 
 import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 
 import 'package:build/build.dart';
@@ -10,6 +11,8 @@ import 'package:built_collection/built_collection.dart';
 import 'package:dart_style/dart_style.dart';
 
 import 'package:source_gen/source_gen.dart';
+// TODO(lejard_h) Code builder not null safe yet
+// ignore: import_of_legacy_library_into_null_safe
 import 'package:code_builder/code_builder.dart';
 import 'package:chopper/chopper.dart' as chopper;
 import 'package:logging/logging.dart';
@@ -68,7 +71,7 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
 
     final friendlyName = element.name;
     final name = '_\$$friendlyName';
-    final baseUrl = annotation?.peek(_baseUrlVar)?.stringValue ?? '';
+    final baseUrl = annotation.peek(_baseUrlVar)?.stringValue ?? '';
 
     final classBuilder = Class((builder) {
       builder
@@ -89,7 +92,7 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
         constructorBuilder.optionalParameters.add(
           Parameter((paramBuilder) {
             paramBuilder.name = _clientVar;
-            paramBuilder.type = refer('${chopper.ChopperClient}');
+            paramBuilder.type = refer('${chopper.ChopperClient}?');
           }),
         );
 
@@ -120,7 +123,7 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
     final parts = _getAnnotations(m, chopper.Part);
     final fileFields = _getAnnotations(m, chopper.PartFile);
 
-    final headers = _generateHeaders(m, method);
+    final headers = _generateHeaders(m, method!);
     final url = _generateUrl(method, paths, baseUrl);
     final responseType = _getResponseType(m.returnType);
     final responseInnerType =
@@ -129,44 +132,34 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
     return Method((b) {
       b.annotations.add(refer('override'));
       b.name = m.displayName;
-      b.returns =
-          Reference(m.returnType.getDisplayString(withNullability: false));
-      b.types.addAll(m.typeParameters
-          .map((t) => Reference(t.getDisplayString(withNullability: false))));
-      b.requiredParameters.addAll(m.parameters
-          .where((p) => p.isNotOptional)
-          .map((p) => Parameter((pb) => pb
-            ..name = p.name
-            ..type =
-                Reference(p.type.getDisplayString(withNullability: false)))));
 
-      b.optionalParameters.addAll(m.parameters
-          .where((p) => p.isOptionalPositional)
-          .map((p) => Parameter((pb) {
-                pb
-                  ..name = p.name
-                  ..type = Reference(
-                      p.type.getDisplayString(withNullability: false));
+      // We don't support returning null Type
+      b.returns = Reference(
+        m.returnType.getDisplayString(withNullability: false),
+      );
 
-                if (p.defaultValueCode != null) {
-                  pb.defaultTo = Code(p.defaultValueCode);
-                }
-                return pb;
-              })));
+      // And null Typed parameters
+      b.types.addAll(
+        m.typeParameters.map(
+          (t) => Reference(t.getDisplayString(withNullability: false)),
+        ),
+      );
+
+      b.requiredParameters.addAll(
+        m.parameters
+            .where((p) => p.isRequiredPositional)
+            .map(buildRequiredPositionalParam),
+      );
 
       b.optionalParameters.addAll(
-          m.parameters.where((p) => p.isNamed).map((p) => Parameter((pb) {
-                pb
-                  ..named = true
-                  ..name = p.name
-                  ..type = Reference(
-                      p.type.getDisplayString(withNullability: false));
+        m.parameters
+            .where((p) => p.isOptionalPositional)
+            .map(buildOptionalPositionalParam),
+      );
 
-                if (p.defaultValueCode != null) {
-                  pb.defaultTo = Code(p.defaultValueCode);
-                }
-                return pb;
-              })));
+      b.optionalParameters.addAll(
+        m.parameters.where((p) => p.isNamed).map(buildNamedParam),
+      );
 
       final blocks = [
         url.assignFinal(_urlVar).statement,
@@ -242,13 +235,13 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
       final requestFactory = factoryConverter?.peek('request');
       if (requestFactory != null) {
         final func = requestFactory.objectValue.toFunctionValue();
-        namedArguments['requestConverter'] = refer(_factoryForFunction(func));
+        namedArguments['requestConverter'] = refer(_factoryForFunction(func!));
       }
 
       final responseFactory = factoryConverter?.peek('response');
       if (responseFactory != null) {
         final func = responseFactory.objectValue.toFunctionValue();
-        namedArguments['responseConverter'] = refer(_factoryForFunction(func));
+        namedArguments['responseConverter'] = refer(_factoryForFunction(func!));
       }
 
       final typeArguments = <Reference>[];
@@ -256,10 +249,10 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
         typeArguments
             .add(refer(responseType.getDisplayString(withNullability: false)));
         typeArguments.add(
-            refer(responseInnerType.getDisplayString(withNullability: false)));
+            refer(responseInnerType!.getDisplayString(withNullability: false)));
       }
 
-      blocks.add(refer('client.send')
+      blocks.add(refer('$_clientVar.send')
           .call([refer(_requestVar)], namedArguments, typeArguments)
           .returned
           .statement);
@@ -270,16 +263,16 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
 
   String _factoryForFunction(FunctionTypedElement function) {
     if (function.enclosingElement is ClassElement) {
-      return '${function.enclosingElement.name}.${function.name}';
+      return '${function.enclosingElement!.name}.${function.name}';
     }
-    return function.name;
+    return function.name!;
   }
 
   Map<String, ConstantReader> _getAnnotation(MethodElement method, Type type) {
     var annotation;
-    String name;
+    var name = '';
     for (final p in method.parameters) {
-      final a = _typeChecker(type).firstAnnotationOf(p);
+      dynamic a = _typeChecker(type).firstAnnotationOf(p);
       if (annotation != null && a != null) {
         throw Exception(
             'Too many $type annotation for \'${method.displayName}\'');
@@ -306,7 +299,7 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
 
   TypeChecker _typeChecker(Type type) => TypeChecker.fromRuntime(type);
 
-  ConstantReader _getMethodAnnotation(MethodElement method) {
+  ConstantReader? _getMethodAnnotation(MethodElement method) {
     for (final type in _methodsAnnotations) {
       final annotation = _typeChecker(type)
           .firstAnnotationOf(method, throwOnUnresolved: false);
@@ -315,7 +308,7 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
     return null;
   }
 
-  ConstantReader _getFactoryConverterAnnotation(MethodElement method) {
+  ConstantReader? _getFactoryConverterAnnotation(MethodElement method) {
     final annotation = _typeChecker(chopper.FactoryConverter)
         .firstAnnotationOf(method, throwOnUnresolved: false);
     if (annotation != null) return ConstantReader(annotation);
@@ -337,19 +330,20 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
     chopper.Patch,
     chopper.Method,
     chopper.Head,
+    chopper.Options,
   ];
 
-  DartType _genericOf(DartType type) {
+  DartType? _genericOf(DartType? type) {
     return type is InterfaceType && type.typeArguments.isNotEmpty
         ? type.typeArguments.first
         : null;
   }
 
-  DartType _getResponseType(DartType type) {
+  DartType? _getResponseType(DartType type) {
     return _genericOf(_genericOf(type));
   }
 
-  DartType _getResponseInnerType(DartType type) {
+  DartType? _getResponseInnerType(DartType type) {
     final generic = _genericOf(type);
 
     if (generic == null ||
@@ -450,9 +444,9 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
         refer(p.displayName),
       ];
 
-      list.add(
-          refer('PartValue<${p.type.getDisplayString(withNullability: false)}>')
-              .newInstance(params));
+      list.add(refer(
+              'PartValue<${p.type.getDisplayString(withNullability: p.type.isNullable)}>')
+          .newInstance(params));
     });
     fileFields.forEach((p, ConstantReader r) {
       final name = r.peek('name')?.stringValue ?? p.displayName;
@@ -469,33 +463,47 @@ class ChopperGenerator extends GeneratorForAnnotation<chopper.ChopperApi> {
     return literalList(list, refer('PartValue'));
   }
 
-  Code _generateHeaders(MethodElement methodElement, ConstantReader method) {
-    final headers = {};
+  Code? _generateHeaders(MethodElement methodElement, ConstantReader method) {
+    final codeBuffer = StringBuffer('')..writeln('{');
 
+    // Search for @Header anotation in method parameters
     final annotations = _getAnnotations(methodElement, chopper.Header);
 
     annotations.forEach((parameter, ConstantReader annotation) {
-      final name =
-          annotation.peek('name')?.stringValue ?? parameter.displayName;
-      headers[literal(name)] = refer(parameter.displayName);
+      final paramName = parameter.displayName;
+      final name = annotation.peek('name')?.stringValue ?? paramName;
+
+      if (parameter.type.isNullable) {
+        codeBuffer.writeln('if ($paramName != null) \'$name\': $paramName,');
+      } else {
+        codeBuffer.writeln('\'$name\': $paramName,');
+      }
     });
 
-    final methodAnnotations = method.peek('headers').mapValue;
+    final headersReader = method.peek('headers');
+    if (headersReader == null) return null;
+
+    final methodAnnotations = headersReader.mapValue;
 
     methodAnnotations.forEach((headerName, headerValue) {
-      headers[literal(headerName.toStringValue())] =
-          literal(headerValue.toStringValue());
+      if (headerName != null && headerValue != null) {
+        codeBuffer.writeln(
+          '\'${headerName.toStringValue()}\': ${literal(headerValue.toStringValue())},',
+        );
+      }
     });
 
-    if (headers.isEmpty) {
+    codeBuffer.writeln('};');
+    final code = codeBuffer.toString();
+    if (code == '{\n};\n') {
       return null;
     }
 
-    return literalMap(headers).assignFinal(_headersVar).statement;
+    return Code('final $_headersVar = $code');
   }
 }
 
-Builder chopperGeneratorFactoryBuilder({String header}) => PartBuilder(
+Builder chopperGeneratorFactoryBuilder({String? header}) => PartBuilder(
       [ChopperGenerator()],
       '.chopper.dart',
       header: header,
@@ -508,3 +516,51 @@ String getMethodPath(ConstantReader method) => method.read('path').stringValue;
 
 String getMethodName(ConstantReader method) =>
     method.read('method').stringValue;
+
+extension DartTypeExtension on DartType {
+  bool get isNullable => nullabilitySuffix != NullabilitySuffix.none;
+}
+
+// All positional required params must support nullability
+Parameter buildRequiredPositionalParam(ParameterElement p) {
+  return Parameter(
+    (pb) => pb
+      ..name = p.name
+      ..type = Reference(
+        p.type.getDisplayString(withNullability: p.type.isNullable),
+      ),
+  );
+}
+
+// All optional positional params must support nullability
+Parameter buildOptionalPositionalParam(ParameterElement p) {
+  return Parameter((pb) {
+    pb
+      ..name = p.name
+      ..type = Reference(
+        p.type.getDisplayString(withNullability: p.type.isNullable),
+      );
+
+    if (p.defaultValueCode != null) {
+      pb.defaultTo = Code(p.defaultValueCode!);
+    }
+  });
+}
+
+// Named params can be optional or required, they also need to support
+// nullability
+Parameter buildNamedParam(ParameterElement p) {
+  return Parameter((pb) {
+    pb
+      ..named = true
+      ..name = p.name
+      ..required = p.isRequiredNamed
+      ..type = Reference(
+        p.type.getDisplayString(withNullability: p.type.isNullable),
+      );
+
+    if (p.defaultValueCode != null) {
+      pb.defaultTo = Code(p.defaultValueCode!);
+    }
+  });
+}